Took methods from NodesApiEndpointsSpec into test-utils and made them more generic
